implement encoding conversion on file open
Without an encoding specified in the file open dialog, xnedit looks in the xattr "charset" and uses the value as encoding. Without an charset xattr, the encoding is detected by looking at the byte order mark (BOM).

Without a BOM, xnedit assumes the file is UTF-8 encoded and just copies the bytes without any conversion.